Fixes to error reporting from registry, and doc fixes.
